+++ src/main/java/com/pedrolg/customer/CustomerDao.java
@@ -6,15 +6,15 @@ import java.util.Optional;
 public interface CustomerDao {
     List<Customer> selectAllCustomers();
 
-    Optional<Customer> selectCustomerById(Integer id);
+    Optional<Customer> selectCustomerById(Long id);
 
     void insertCustomer(Customer customer);
 
-    void deleteCustomerById(Integer id);
+    void deleteCustomerById(Long id);
 
     boolean existsPersonByEmail(String email);
 
-    boolean existsPersonWithId(Integer id);
+    boolean existsPersonWithId(Long id);
 
     void updateCustomer(Customer customer);
 }
