+++ src/main/java/com/pedrolg/customer/CustomerJDBCDataAccessService.java
@@ -29,7 +29,7 @@ public class CustomerJDBCDataAccessService implements CustomerDao {
     }
 
     @Override
-    public Optional<Customer> selectCustomerById(Integer id) {
+    public Optional<Customer> selectCustomerById(Long id) {
         var sql = """
                 SELECT id, name, email, age
                 FROM customer
@@ -51,7 +51,7 @@ public class CustomerJDBCDataAccessService implements CustomerDao {
     }
 
     @Override
-    public void deleteCustomerById(Integer id) {
+    public void deleteCustomerById(Long id) {
         var sql = """
                 DELETE FROM customer
                 WHERE id= ?;
@@ -72,7 +72,7 @@ public class CustomerJDBCDataAccessService implements CustomerDao {
     }
 
     @Override
-    public boolean existsPersonWithId(Integer id) {
+    public boolean existsPersonWithId(Long id) {
         String sql = """
                 SELECT count(id)
                 FROM customer
